refactor : 회원 탈퇴 메소드가 많아서 클래스 추출
issue : #53
